Moved score-stuff from aibrain to lua/score.lua
