+++ lua/aibrain.lua
@@ -30,27 +30,13 @@ local AIAttackUtils = import('/lua/AI/aiattackutilities.lua')
 local SUtils = import('/lua/AI/sorianutilities.lua')
 local StratManager = import('/lua/sim/StrategyManager.lua')
 
-local scoreOption = ScenarioInfo.Options.Score or "no"
+
 
 local TransferUnitsOwnership = import('/lua/SimUtils.lua').TransferUnitsOwnership
+local CalculateBrainScore = import('/lua/score.lua').CalculateBrainScore
 
 
 local observer = false
-scoreData = {}
-scoreData.current = {}
-
-scoreInterval = 10
-scoreData.historical = {}
--- copy data over to historical
-local curInterval = 1
-
-local historicalUpdateThread = ForkThread(function()
-    while true do
-        WaitSeconds(scoreInterval)
-        scoreData.historical[curInterval] = table.deepcopy(scoreData.current)
-        curInterval = curInterval + 1
-    end
-end)
 
 ----end sorian ai imports
 
@@ -69,276 +55,6 @@ local VOReplayTime = {
     CoordinatedAttackInitiated = 5,
     NuclearLaunchDetected = 1,
 }
-
-------------------------------------------------------------------------------------------
------------- Runtime score update sync loop  ------------
-------------------------------------------------------------------------------------------
-local ArmyScore = {}
-
-function UpdateScoreData(newData)
-    scoreData.current = table.deepcopy(newData)
-    fullSyncOccured = false
-end
-
-
-function StopScoreUpdate()
-    if historicalUpdateThread then
-        KillThread(historicalUpdateThread)
-    end
-end
-
-
-function CollectCurrentScores()
-    -- Initialize the score data stucture
-    for index, brain in ArmyBrains do
-        ArmyScore[index] = {
-            -- General scores
-            general = {
-                score = 0,
-                mass = 0,
-                energy = 0,
-                kills = {
-                    count = 0,
-                    mass = 0,
-                    energy = 0
-                },
-                built = {
-                    count = 0,
-                    mass = 0,
-                    energy = 0
-                },
-                lost = {
-                    count = 0,
-                    mass = 0,
-                    energy = 0
-                },
-                currentunits = {
-                    count = 0
-                },
-                currentcap = {
-                    count = 0
-                }
-            },
-
-            -- Unit scores
-            units = {
-                cdr = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                },
-                land = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                },
-                air = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                },
-                naval = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                },
-                structures = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                },
-                experimental = {
-                    kills = 0,
-                    built = 0,
-                    lost = 0
-                }
-            },
-
-            -- Resource scores
-            resources = {
-                massin = {
-                    total = 0,
-                    rate = 0
-                },
-                massout = {
-                    total = 0,
-                    rate = 0
-                },
-                energyin = {
-                    total = 0,
-                    rate = 0
-                },
-                energyout = {
-                    total = 0,
-                    rate = 0
-                },
-                massover = 0,
-                energyover = 0
-            }
-        }
-    end
-    UpdateScoreData(ArmyScore)
-
-    -- Collect the various scores at regular intervals
-    while true do
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------------
-           ---- General economy scores 1 ----
-           ------------------------------------------------------------
-           ArmyScore[index].general.score = brain:CalculateScore()
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------------
-           ---- General economy scores 2 ----
-           ------------------------------------------------------------
-           ArmyScore[index].general.mass = brain:GetArmyStat("Economy_TotalProduced_Mass", 0.0).Value
-           ArmyScore[index].general.energy = brain:GetArmyStat("Economy_TotalProduced_Energy", 0.0).Value
-           ArmyScore[index].general.currentunits.count = brain:GetArmyStat("UnitCap_Current", 0.0).Value
-           ArmyScore[index].general.currentcap.count = brain:GetArmyStat("UnitCap_MaxCap", 0.0).Value
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------------------
-           ---- General unit stats scores 1 ----
-           ------------------------------------------------------------------
-           ArmyScore[index].general.kills.count = brain:GetArmyStat("Enemies_Killed", 0.0).Value
-           ArmyScore[index].general.kills.mass = brain:GetArmyStat("Enemies_MassValue_Destroyed", 0.0).Value
-           ArmyScore[index].general.kills.energy = brain:GetArmyStat("Enemies_EnergyValue_Destroyed", 0.0).Value
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------------------
-           ---- General unit stats scores 2 ----
-           ------------------------------------------------------------------
-           ArmyScore[index].general.built.count = brain:GetArmyStat("Units_History", 0.0).Value
-           ArmyScore[index].general.built.mass = brain:GetArmyStat("Units_MassValue_Built", 0.0).Value
-           ArmyScore[index].general.built.energy = brain:GetArmyStat("Units_EnergyValue_Built", 0.0).Value
-           ArmyScore[index].general.lost.count = brain:GetArmyStat("Units_Killed", 0.0).Value
-           ArmyScore[index].general.lost.mass = brain:GetArmyStat("Units_MassValue_Lost", 0.0).Value
-           ArmyScore[index].general.lost.energy = brain:GetArmyStat("Units_EnergyValue_Lost", 0.0).Value
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------
-           ---- Regular unit scores 1 ----
-           ------------------------------------------------------
-           ArmyScore[index].units.land.kills = brain:GetBlueprintStat("Enemies_Killed", categories.LAND)
-           ArmyScore[index].units.land.built = brain:GetBlueprintStat("Units_History", categories.LAND)
-           ArmyScore[index].units.land.lost = brain:GetBlueprintStat("Units_Killed", categories.LAND)
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ------------------------------------------------------
-           ---- Regular unit scores 2 ----
-           ------------------------------------------------------
-           ArmyScore[index].units.air.kills = brain:GetBlueprintStat("Enemies_Killed", categories.AIR)
-           ArmyScore[index].units.air.built = brain:GetBlueprintStat("Units_History", categories.AIR)
-           ArmyScore[index].units.air.lost = brain:GetBlueprintStat("Units_Killed", categories.AIR)
-           ArmyScore[index].units.naval.kills = brain:GetBlueprintStat("Enemies_Killed", categories.NAVAL)
-           ArmyScore[index].units.naval.built = brain:GetBlueprintStat("Units_History", categories.NAVAL)
-           ArmyScore[index].units.naval.lost = brain:GetBlueprintStat("Units_Killed", categories.NAVAL)
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ----------------------------------------------------------------------------------
-           ---- Structures and Special units scores ----
-           ----------------------------------------------------------------------------------
-           ArmyScore[index].units.cdr.kills = brain:GetBlueprintStat("Enemies_Killed", categories.COMMAND)
-           ArmyScore[index].units.cdr.built = brain:GetBlueprintStat("Units_History", categories.COMMAND)
-           ArmyScore[index].units.cdr.lost = brain:GetBlueprintStat("Units_Killed", categories.COMMAND)
-           ArmyScore[index].units.experimental.kills = brain:GetBlueprintStat("Enemies_Killed", categories.EXPERIMENTAL)
-           ArmyScore[index].units.experimental.built = brain:GetBlueprintStat("Units_History", categories.EXPERIMENTAL)
-           ArmyScore[index].units.experimental.lost = brain:GetBlueprintStat("Units_Killed", categories.EXPERIMENTAL)
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ----------------------------------------------
-           ---- Structures scores ----
-           ----------------------------------------------
-           ArmyScore[index].units.structures.kills = brain:GetBlueprintStat("Enemies_Killed", categories.STRUCTURE)
-           ArmyScore[index].units.structures.built = brain:GetBlueprintStat("Units_History", categories.STRUCTURE)
-           ArmyScore[index].units.structures.lost = brain:GetBlueprintStat("Units_Killed", categories.STRUCTURE)
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ----------------------------------------------
-           ---- Resource scores 1 ----
-           ----------------------------------------------
-           ArmyScore[index].resources.massin.total = brain:GetArmyStat("Economy_TotalProduced_Mass", 0.0).Value
-           ArmyScore[index].resources.massin.rate = brain:GetArmyStat("Economy_Income_Mass", 0.0).Value
-           ArmyScore[index].resources.massout.total = brain:GetArmyStat("Economy_TotalConsumed_Mass", 0.0).Value
-           ArmyScore[index].resources.massout.rate = brain:GetArmyStat("Economy_Output_Mass", 0.0).Value
-           ArmyScore[index].resources.massover = brain:GetArmyStat("Economy_AccumExcess_Mass", 0.0).Value
-        end
-        WaitSeconds(0.5)  -- update scores every second
-
-        for index, brain in ArmyBrains do
-           ----------------------------------------------
-           ---- Resource scores 2 ----
-           ----------------------------------------------
-           ArmyScore[index].resources.energyin.total = brain:GetArmyStat("Economy_TotalProduced_Energy", 0.0).Value
-           ArmyScore[index].resources.energyin.rate = brain:GetArmyStat("Economy_Income_Energy", 0.0).Value
-           ArmyScore[index].resources.energyout.total = brain:GetArmyStat("Economy_TotalConsumed_Energy", 0.0).Value
-           ArmyScore[index].resources.energyout.rate = brain:GetArmyStat("Economy_Output_Energy", 0.0).Value
-           ArmyScore[index].resources.energyover = brain:GetArmyStat("Economy_AccumExcess_Energy", 0.0).Value
-        end
-        WaitSeconds(0.5)  -- update scores every second
-        UpdateScoreData(ArmyScore)
-    end
-
-end
-
-function SyncScores()
-    if GetFocusArmy() == -1 or import('/lua/victory.lua').gameOver == true or observer == true then
-        observer = true
-        Sync.FullScoreSync = true
-        Sync.ScoreAccum = scoreData
-        Sync.Score = scoreData.current
-
-    elseif observer == false then
-        for index, brain in ArmyBrains do
-            Sync.Score[index] = {}
-            Sync.Score[index].general = {}
-            if GetFocusArmy() == index or GetFocusArmy() == -1 then
-                Sync.Score[index].general.currentunits = {}
-                Sync.Score[index].general.currentunits.count = ArmyScore[index].general.currentunits.count
-                Sync.Score[index].general.currentcap = {}
-                Sync.Score[index].general.currentcap.count = ArmyScore[index].general.currentcap.count
-            end
-
-            ----------------------------------------
-            ---- General scores ----
-            ----------------------------------------
-            if scoreOption ~= 'no' then
-                Sync.Score[index].general.score = ArmyScore[index].general.score
-            else
-                Sync.Score[index].general.score = -1
-            end
-        end
-
-    end
-
-
-end
-
-function SyncCurrentScores()
-    Sync.FullScoreSync = false
-    -- Sync the score at 1 sec intervals
-    while true do
-        SyncScores()
-        WaitSeconds(1)  -- update scores every second
-    end
-end
 
 AIBrain = Class(moho.aibrain_methods) {
 
@@ -606,35 +322,7 @@ AIBrain = Class(moho.aibrain_methods) {
         return false
     end,
 
-    CalculateScore = function(self)
-        local commanderKills = self:GetArmyStat("Enemies_Commanders_Destroyed",0).Value
-        local massSpent = self:GetArmyStat("Economy_TotalConsumed_Mass",0.0).Value
-        local massProduced = self:GetArmyStat("Economy_TotalProduced_Mass",0.0).Value -- not currently being used
-        local energySpent = self:GetArmyStat("Economy_TotalConsumed_Energy",0.0).Value
-        local energyProduced = self:GetArmyStat("Economy_TotalProduced_Energy",0.0).Value -- not currently being used
-        local massValueDestroyed = self:GetArmyStat("Enemies_MassValue_Destroyed",0.0).Value
-        local massValueLost = self:GetArmyStat("Units_MassValue_Lost",0.0).Value
-        local energyValueDestroyed = self:GetArmyStat("Enemies_EnergyValue_Destroyed",0.0).Value
-        local energyValueLost = self:GetArmyStat("Units_EnergyValue_Lost",0.0).Value
-
-        -- helper variables to make equation more clear
-        local excessMassProduced = massProduced - massSpent -- not currently being used
-        local excessEnergyProduced = energyProduced - energySpent -- not currently being used
-        local energyValueCoefficient = 20
-        local commanderKillBonus = commanderKills + 1 -- not currently being used
-
-        -- score components calculated
-        local resourceProduction = ((massSpent) + (energySpent / energyValueCoefficient)) / 2
-        local battleResults = (((massValueDestroyed - massValueLost- (commanderKills * 18000)) + ((energyValueDestroyed - energyValueLost - (commanderKills * 5000000)) / energyValueCoefficient)) / 2)
-        if battleResults < 0 then
-            battleResults = 0
-        end
-
-        -- score calculated
-        local score = math.floor(resourceProduction + battleResults + (commanderKills * 5000))
-
-        return score
-    end,
+    
 
     ------------------------------------------------------------------------------------------------------------------------------------------
     ---- ------------- TRIGGERS BASED ON AN AI BRAIN       ------------- ----
@@ -892,7 +580,7 @@ AIBrain = Class(moho.aibrain_methods) {
                 -- this part determines who the allies are
                 for index, brain in ArmyBrains do
                     brain.index = index
-                    brain.score = brain:CalculateScore()
+                    brain.score = CalculateBrainScore(brain)
                     if IsAlly(selfIndex, brain:GetArmyIndex()) and selfIndex ~= brain:GetArmyIndex() and not brain:IsDefeated() then
                         table.insert(allies, brain)
                     end
